Clean-up LanguageClient
Clean-up LanguageClient class so it is easier to work on it (might do multiple standalone changes so they are easier to debug/revert).
